building out add property

+++ src/main/java/com/pryhmez/collabomain/properties/PropertyRepository.java
@@ -2,6 +2,6 @@ package com.pryhmez.collabomain.properties;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PropertyRepository extends JpaRepository<Property, Long> {
+public interface PropertyRepository extends JpaRepository<Property, Integer> {
 
 }
